+++ include/crnum.hpp
@@ -20,7 +20,7 @@ class CRnum final: public CRobj {
         double valueof() const;
         bool isnumber() const override;
         void print_tree() const override;
-        std::string genCode(size_t parent, size_t index, ssize_t place,std::string indent) const override;
+        std::string genCode(size_t parent, size_t index, int place,std::string indent) const override;
 
         void shift(size_t index) override final {
             return;
