+++ src/_ast.cpp
@@ -360,7 +360,7 @@ void ASTnode::_creval()
         }
 
         // 3) advance the odometer
-        ssize_t i = paramsize;
+        int i = paramsize;
         while (i >= 0) {
             ind[i]++;
             if (ind[i] < params[i]) {
